PEAR-1860 - convert image viewer slices
