+++ packages/core/src/features/imageDetails/index.ts
@@ -1,0 +1,4 @@
+import { useImageViewerQuery } from "./imageViewer";
+import { useImageDetailsQuery } from "./imageDetailsSlice";
+
+export { useImageViewerQuery, useImageDetailsQuery };
