+++ packages/core/src/index.ts
@@ -8,8 +8,7 @@ export * from "./features/gdcapi/gdcapi";
 export * from "./features/gdcapi/filters";
 export * from "./features/gdcapi/gdcgraphql";
 export * from "./features/gdcapi/types";
-export * from "./features/imageDetails/imageDetailsSlice";
-export * from "./features/imageDetails/imageViewer";
+export * from "./features/imageDetails";
 export * from "./features/facets/facetSlice";
 export * from "./features/facets/types";
 export * from "./features/facets/facetApiGQL";
